// packages/utils/types/tweet.ts
export interface Media {
  display_url: string
  expanded_url: string
  id_str: string
  indices: number[]
  media_url_https: string
  type: 'video' | 'photo' | 'animated_gif'
  additional_media_info?: {
    title: string
    description: string
  }
  mediaStats?: {
    viewCount: number
  }
  url: string
  features: {
    large: Feature
    medium: Feature
    small: Feature
    orig: Feature
  }
  sizes: {
    large: Size
    medium: Size
    small: Size
    thumb: Size
  }
  original_info: {
    height: number
    width: number
    focus_rects?: FocusRect[]
  }
  video_info?: {
    aspect_ratio: number[]
    duration_millis: number
    variants: Variant[]
  }
  ext_alt_text?: string
  media_key?: string
  ext_media_availability?: {
    status: string
  }
}

export interface Variant {
  bitrate?: number
  content_type: string
  url: string
}

interface Feature {
  faces: FocusRect[]
}

interface FocusRect {
  x: number
  y: number
  h: number
  w: number
}

interface Size {
  h: number
  w: number
  resize: string
}